[BUGFIX] Wait for remote repository to be updated
If local changes happen during the release process, e.g. for
modifying version number and extracting ChangeLog data, a local
commit is generated. The commit hash locally does not match with
the commit hash that is actually cherry-picked by Gerrit to the
remote repository.
To have valid tags, it's required to wait and ensure the remote
repository is updated - since there's synchronisation latency
between review.typo3.org and git.typo3.org as well.

Change-Id: I48d249368da3aec914d7a6925c3ca965d178caa7
Reviewed-on: https://review.typo3.org/20088
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader